Added an optional checkPermission function and upgraded quick.db.

+++ commands/Owners Only/poweroff.js
@@ -8,16 +8,16 @@
  * *************************************/
 
 exports.run = async (bot, msg) => {
-  if (!['242775871059001344', '247221105515823104', '236279900728721409'].includes(msg.author.id)) {
-    msg.reply('Nope! You need the person who created this bot to use this command.');
-  }
-  else {
-    await msg.reply(':warning: Pokebot is now powering off!');
-    await bot.user.setStatus('invisible')
-    console.log('Pokebot is now powering off...');
-    process.exit(0);
-  }
+  await msg.reply(':warning: Pokebot is now powering off!');
+  await bot.user.setStatus('invisible')
+  console.log('Pokebot is now powering off...');
+  process.exit(0);
 };
+
+exports.checkPermission = (bot, member) => {
+  if (!['242775871059001344', '247221105515823104', '236279900728721409'].includes(member.id)) return false;
+  return true;
+}
 
 exports.conf = {
   aliases: ['reboot', 'restart'],
